facebook-insights-microservice/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
motor==2.5.1
pydantic==1.8.2
aioredis==2.0.0
aioboto3==9.3.0
selenium==4.1.0
beautifulsoup4==4.10.0
python-dotenv==0.19.1
loguru==0.5.3

